requirements.txt
gradio
python-dotenv
seaborn
pandas
openai